+++ CHANGELOG.md
@@ -6,6 +6,41 @@
 - Continued development to encompass any new documented features of the CyberArk API.
 - psPAS v5.0...
 
+## **4.1.X** (July X 2020)
+
+### Module update to cover CyberArk 11.5 API features
+
+- **Behavior Changes**
+  - `Get-PASPlatform`
+    - When invoked with no parameters to return details of all configured platforms, defaults to operation against the endpoint for the 11.4 API.
+    - When invoked with a value provided for the `Active` parameter, will perform operation against the endpoint for the 11.4 API.
+    - To utilise the 11.1 api endpoint, a value should be provided for the `PlatformType` and/or `Search` parameters,  or, `Active` and `PlatformType` and/or `Search` parameters.
+
+- New Functions
+  - `Copy-PASPlatform`
+    - Duplicates target, dependent, group or rotational group platform to a new platform.
+    - 11.4 functionality, missed in the `4.0.0` release.
+  - `Disable-PASPlatform`
+    - Disables, target, group or rotational group platform.
+    - 11.4 functionality, missed in the `4.0.0` release.
+  - `Enable-PASPlatform`
+    - Enables, target, group or rotational group platform.
+    - 11.4 functionality, missed in the `4.0.0` release.
+  - `Remove-PASPlatform`
+    - Deletes, target, dependent, group or rotational group platform.
+    - 11.4 functionality, missed in the `4.0.0` release.
+  - `Remove-PASGroup`
+    - Deletes a specified vault user group
+    - 11.5 functionality.
+
+- Updated Functions
+  - `Get-PASUser`
+    - 11.5 output includes group membership details.
+    - group membership property may be included in output when function is executed from earlier versions, but its content will be blank.
+
+- Other Fixes & Updates
+  - Documentation updated.
+
 ## **4.0.0** (July 1st 2020)
 
 ### Module update to cover CyberArk 11.4 API features